Implement Signature Scanning (#6)
This provides an auto splitter side implementation of signature
scanning. It's unclear still to what degree the runtime should support
signature scanning, but for a few use cases an auto splitter side
implementation works fine for now and lets us experiment a little in
terms of performance, what the API looks like, what algorithms to use,
whether it needs to be done in the runtime and if so, if it should be
asynchronous to not hold up the `update` function.