Support Floating Point Number Formatting (#9)
Auto Splitters written in Rust tend to be a lot smaller when not using
`std` or even `core::fmt`. Using crates that directly format numbers
results in much smaller binaries. We previously added the `itoa` crate
for formatting integers. This brings in `ryu` which is almost identical,
but works on floating point numbers. However by default it brings in a
fairly large lookup table, so there's an alterntive feature that can be
activated to trade off size for speed.