Relax the `bytemuck` bounds
It's actually fine if the structs we read have padding, there's no
reason why that would cause any Undefined Behavior. So we can use
`AnyBitPattern` instead as the bound. However, I also noticed that
`bytemuck` has `CheckedBitPattern` which builds on top of
`AnyBitPattern` but allows for types where not every bit pattern is
valid. For those `bytemuck` then defines a function that allows you to
check the validity. `CheckedBitPattern` is implemented for
`AnyBitPattern` (by doing nothing), so we can just always use
`CheckedBitPattern` and support even more complex types.