Implement a Mutex for single threaded WebAssembly (#22)
WebAssembly usually is single threaded, which means that a simple
`RefCell` is sufficient for ensuring you have mutable access to global
data. The `Mutex` in `std` works the same way on single threaded
WebAssembly, however your auto splitter may not use `std`. So we provide
essentially the same implementation, but it works on `no_std`.