Add a macro for defining a panic handler
This adds a macro `asr::panic_handler!()` which allows the user to
easily define a panic handler for their auto splitter. By default it
prints a message of up to 1024 bytes in debug builds and doesn't print
anything in release builds. This is all highly configurable.

```rust
asr::panic_handler! {
    /// When to print the panic message.
    /// Default: debug
    print: never | debug | always,

    /// The size of the stack based buffer in bytes.
    /// Default: 1024
    buffer: <number>,
}
```