Allow accessing the path of a process (#37)
With the newly added WASI support, you may want to read the actual
executable of a process or files surrounding the executable. It is
therefore important that you can get the path of the process to locate
its executable in the file system. This adds the functionality to do so.
However, because paths are potentially unbounded in size we gate this
functionality behind a new `alloc` feature.