Add Tooltips and Titles to the User Settings (#45)
Auto Splitters can now specify tooltips for the user settings.
Additionally there's a new type of user setting. Titles allow users to
group settings together and give them a name. This is useful for
settings that are related to each other. Heading levels can also be
specified to create a hierarchy of settings. Unlike ASL, the titles do
not by themselves have a value. A user interface could however still
allow toggling the grouped settings by clicking on the title.