Replacing `ArrayVec` in `DeepPointer` with standard arrays
This replaces the internal `ArrayVec` in the `DeepPointer` struct with a standard array.

This change is necessary for 2 reasons:
1. The features provided by ArrayVec are not necessary for this implementation of DeepPointer
2. Using a standard array allows to use the `Copy` trait on the `DeepPointer` struct, which is something ArrayVec is preventing.

Additionally, the `assert!` has been now removed. If a new `DeepPointer` is created with a shorter `CAP` than the number of offsets, the pointer path will just be truncated according to the value of `CAP`.