Add Support for the Godot Engine
This adds support for querying information and navigating the node tree
in games built with the Godot Engine. This is initial version only
supports games using Godot 4.2 without debug symbols and is missing many
features such as accessing variables inside scripts.

It is intentionally structured and documented in a way to match both the
Godot documentation and the source code folder structure.